+++ pydoclint/utils/return_arg.py
@@ -1,0 +1,10 @@
+from dataclasses import dataclass
+
+
+@dataclass
+class ReturnArg:
+    """A class to hold one return argument in the docstring's return section"""
+
+    argName: str
+    argType: str
+    argDescr: str