+++ tests/data/google/returns/function.py
@@ -85,14 +85,18 @@ def func5() -> int:
 
 def func6():
     """
+    If no summary here, the parser will mis-parse the return section
+
     Returns:
         int: Something to return
     """
     print(123)
 
 
-def func6() -> float:
+def func62() -> float:
     """
+    If no summary here, the parser will mis-parse the return section
+
     Returns:
         int: Something to return
     """
@@ -101,7 +105,30 @@ def func6() -> float:
 
 def func7():
     """
+    If no summary here, the parser will mis-parse the return section
+
     Returns:
         int: Something to return
     """
     return 123
+
+
+def func81() -> Tuple[int, bool]:
+    """
+    If no summary here, the parser will mis-parse the return section
+
+    Returns:
+        Tuple[int, bool]: Something to return
+    """
+    return (1, 1.1)
+
+
+def func82() -> Tuple[int, bool]:
+    """
+    If no summary here, the parser will mis-parse the return section
+
+    Returns:
+        int: Integer to return
+        bool: Boolean to return
+    """
+    return (1, 1.1)