+++ tests/data/numpy/returns/method.py
@@ -42,7 +42,7 @@ class MyClass:
 
         Returns
         -------
-        int :
+        int
             Result
         """
         return 1
@@ -53,7 +53,7 @@ class MyClass:
 
         Returns
         -------
-        int :
+        int
             Result
         """
         return 1
@@ -64,7 +64,7 @@ class MyClass:
 
         Returns
         -------
-        float :
+        float
             Result
         """
         return 1.0
@@ -75,7 +75,7 @@ class MyClass:
 
         Returns
         -------
-        int :
+        int
             Result
         """
 
@@ -94,16 +94,16 @@ class MyClass:
         """
         Returns
         -------
-        int:
+        int
             Something to return
         """
         print(123)
 
-    def func6(self) -> float:
+    def func62(self) -> float:
         """
         Returns
         -------
-        int:
+        int
             Something to return
         """
         print(123)
@@ -112,7 +112,31 @@ class MyClass:
         """
         Returns
         -------
-        int:
+        int
             Something to return
         """
         return 123
+
+    def func81(self) -> Tuple[int, bool]:
+        """
+        If no summary here, the parser will mis-parse the return section
+
+        Returns
+        -------
+        Tuple[int, bool]
+            Something to return
+        """
+        return (1, 1.1)
+
+    def func82(self) -> Tuple[int, bool]:
+        """
+        If no summary here, the parser will mis-parse the return section
+
+        Returns
+        -------
+        int
+            Integer to return
+        bool
+            Boolean to return
+        """
+        return (1, 1.1)